+++ llm_generate/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import llm_thread
